Add unified parallel test runner and refactor spec tests
- Add shared.run_parallel_tests helper with thread-safe output capture and interrupt handling
- Refactor spec tests to use shared.run_parallel_tests

The eventual goal here is to have all the suites using a ThreadPool to
parallelize more.
